==> code/01_Assemble_Network_Dataframe.Rmd <==
---
title: "01_Assemble_Network_Dataframe"
output: html_document
---

```{r setup, include=FALSE}
setwd('~/igraph_Network_Expansion/')

knitr::opts_chunk$set(echo = TRUE)
setwd('~/igraph_Network_Expansion/')
library(igraph)
library(paxtoolsr)
library(RJSONIO)
library(httr)
library(DOSE)
library(org.Hs.eg.db)
library(clusterProfiler)
library(parallel)
library(doParallel)
library(RColorBrewer)
#devtools::install_github("jhrcook/HotNetvieweR")
library(HotNetvieweR)
source('igraphhack/igraphplot2.R')
```

## R Import Data

```{r Import}
setwd('~/igraph_Network_Expansion/')

reticulate::use_python("/usr/bin/python", required = TRUE)
[synapseclient <- reticulate::import("synapseclient")
syn_temp <- synapseclient$Synapse()
syn_temp$login(  )

SIFs <- c('syn21914063',
  'syn21914056',
  'syn21914057',
  'syn21914059',
  'syn21914061',
  'syn21914064',
  'syn21914065',
  'syn21914070',
  'syn21914071',
  'syn21914072',
  'syn21914078',
  'syn21914082',
  'syn21914083',
  'syn21914087',
  'syn21914090',
  'syn21914091',
  'syn21914093',
  'syn21914094',
  'syn21914074'
)

names(SIFs) <- c( 'Detailed', 
                  'bind', 
                  'biogrid', 
                  'corum', 
                  'ctd',
                  'dip',
                  'drugbank',
                  'hprd',
                  'humancyc',
                  'intacsyngnc',
                  'msigdb',
                  'netpath',
                  'panther',
                  'pathbank',
                  'pid',
                  'psp',
                  'reactome',
                  'reconx',
                  'kegg'
  )

#Load and Add names: SIFs
boot <- function( i ){
  foo <- read.table( file=syn_temp$get( SIFs[i] )$path , header = F, sep='\t' )
  foo$Pathway <- as.character( i ) 
  return(foo)
}

Loader <- function( SIFs){
  #'@SIFs An object that has SynID values and named for the database
  ##  Each Syn ID is a SIF format file 
  
  total <- list()
  Trial <- for( i in names(SIFs) ){
    temp <- boot( i )
    #eval(parse(text=paste0( 'total$', as.character(NAMES[i]), ' <- temp' )))
    total[[i]] <- temp
  }
  
  Total <- do.call(rbind, total)
  Total <- Total[ , c("V1", "V3", "V2", "Pathway") ]
  colnames(Total) <- c("from", "to", "interaction", "pathway")
  
  #Caluclate how many times this intereactcion was found in all databases:
  Total$Occurance <- paste0( Total$from, '-', Total$to, ':', Total$interaction )
  foo <- paste0( Total$from, '-', Total$to, ':', Total$interaction )
  TabFoo <- table(foo)
  Total$Occurance <-  as.numeric( TabFoo[ Total$Occurance ] )
  
  Genes <- c( as.character(Total$from), as.character(Total$to) )
  Genes <- Genes[ !duplicated(Genes) ]
  Genes <- as.data.frame( Genes )
  
  #Make the pathway column into a list object
  #library(dplyr)
  #library(reshape)
  
  library(data.table)
  Total$UniqCol <- paste0( as.character(Total$from),':', as.character(Total$to),'-', as.character(Total$interaction) )
  dt <- data.table(Total[, c('UniqCol','pathway')])
  DT <- dt[,lapply(.SD, function(col) paste(col, collapse=", ")), by=.(UniqCol)]
  
  sinl<-DT
  
  foo <- as.data.frame( DT )
  poo <- as.list( strsplit(as.character(foo$pathway),',') )
  names(poo) <- foo$UniqCol
  
  totals <- Total[ !duplicated(Total$UniqCol), ]
  pathways <- poo
  
  table(names(pathways) == as.character(totals$UniqCol))
  table( as.character(totals$UniqCol) == names(pathways) )
  
  totals$PATH <- pathways
  totals$PATH <- lapply( totals$PATH,
                 function(x) gsub(" ","", x)
               )
  
  Total <- totals[,c("from", "to", "interaction", "Occurance", "UniqCol", "PATH")]
  colnames(Total) <- c("from", "to", "interaction", "Occurance", "UniqCol", "pathway")
  
  #Make into a network graph
  #Make into a network graph
  
  Graph <- list()
  for(type in levels(Total$interaction) ){
    #NET <- graph_from_data_frame(d=Total[ Total$interaction == type, ], vertices=Genes, directed=T) 
    #set_graph_attr(NET,"interaction","controls-transport-of-chemical")
    eval( parse( text=paste0('Graph$`', type, '` <- graph_from_data_frame(d=Total[ Total$interaction == \'', type, '\', ], vertices=Genes, directed=T) ') ) )
  }
  #Edge Sum is 1903389
  
  #Another way to make the Network object But has multiple edges per-Vertex set
  net_oldStyle <- graph_from_data_frame(d=Total, vertices=Genes, directed=T) 
  return( list( DF=Total, NET=net_oldStyle, Gra=Graph ) )
}

Data <- Loader(SIFs)
Total <- Data$DF
net_oldStyle <- Data$NET
Graph <- Data$Gra

########################################
# write total Gene Space for Jesse
# Genes in pathway commons - No CHEBI
foo <- Total[ !( grepl('CHEBI',Total$from)| grepl('CHEBI',Total$from) ), ]
val.genes <- c(foo$from,foo$to)
val.genes <- val.genes[!duplicated(val.genes)]

write.csv( val.genes, file = 'Non_CHEBI_Connected_Pathway_Commons_Genes.csv' )


parentId <- 'syn26844059'
activityName = 'All Pathway Commons Genes excluding CHEBI Links';
activityDescription = 'All Pathway Commons Genes excluding CHEBI Links';
CODE <- syn_temp$store(synapseclient$Folder(name = "igraphData", parentId = parentId))

thisFileName <- '01_Assemble_Network_Dataframe.Rmd'

# Github link
thisRepo <- githubr::getRepo(repository = "jgockley62/igraph_Network_Expansion", ref="branch", refName='master')
thisFile <- githubr::getPermlink(repository = thisRepo, repositoryPath=paste0('code/',thisFileName))

#Set Used SynIDs For Provenance
Syns_Used <- SIFs

# Set annotations
all.annotations = list(
  dataType = 'Network',
  summaryLevel = 'gene',
  assay	 = 'RNAseq',
  tissueTypeAbrv	= c('IFG', 'STG', 'FP', 'PHG', 'TCX', 'DLFPC'), 
  study = c( 'MSBB', 'ROSMAP', 'Mayo' ), 
  organism = 'HomoSapiens',
  consortium	= 'TreatAD',
  genomeAssemblyID = 'GRCh38'
)

# Store  executed = thisFile,
ENRICH_OBJ <-  syn_temp$store( synapseclient$File( path='01_Assemble_Network_Dataframe.Rmd', name = 'Genes In Pathway Commons', parentId=CODE$properties$id ), used = Syns_Used, executed = thisFile, activityName = activityName, activityDescription = activityDescription)

########################################

length(all.genes[!grepl('CHEBI',all.genes)])
# 19572
length(val.genes[!grepl('CHEBI',val.genes)])
# 19092

Tot <- 0
for(type in levels(Total$interaction) ){
  message( paste0( type, ": ", eval(parse(text=paste0(  'length(E(Graph$`', type, '`))' ))) ))
  Tot <- eval(parse(text=paste0( 'length(E(Graph$`', type, '`))') ))+Tot
}
message( paste0( "Total Interactions = ", Tot ) )


#_# Test Code
#@sk:All interactions of our interested partners
Total[ (Total$from =='PIK3R1' & Total$to=='PIK3CA') | (Total$from =='PIK3CA' & Total$to=='PIK3R1') , ]

#@sk:Do we have multiple edge IDs:
table( which_multiple(net_oldStyle, eids = E(net_oldStyle)) )
#   FALSE    TRUE 
# 1804977   98412
#---# Yes We Do

#@sk:
fun<-Total
fun$Extra <- paste0( as.character(fun$from), ':', as.character(fun$to) )
table(table(fun$Extra))

#Get the shortest path...
remove <- names(V(net_oldStyle))[ grepl( 'CHEBI:', names(V(net_oldStyle)) ) ==T ]
Slim <- delete_vertices( net_oldStyle, remove )

news.path <- shortest_paths( Slim, 
                                from = V(Slim)["SMOC1"], 
                                to  = V(Slim)["SORL1"],
                                mode = "all",
                                output = c('both')
                            )# both path nodes and edges

Trial.path <- induced_subgraph( net_oldStyle, colnames(do.call( rbind, news.path$vpath )) )
igraph::plot.igraph(Trial.path, edge.arrow.size=.4)

# Get the vertecies of N Gene Ids
####################################################################################
MoesinScores <- data.table::fread(syn_temp$get('syn21652325')$path)
MoesinGenes <- MoesinScores$GeneName
GeneSetA <- c( "A1BG", "ABCC5", "ABI1", "ABI1", "ABI2", "ACACA", "PIK3R1", "PIK3CA" )
GeneSetB <- c("CD44", "FLT1", "NTN1", "SFRP1", "STX4", "LACTB", "PVR", "ACE", "EPHX2", "RABEP1", "MADD", "APP", "MAPT")

#Subgraph:
Trial.path <- induced_subgraph( net_oldStyle, GeneSetB )

GS <- GeneSetB
message( paste0( length( GS[ GS%in% names(V(net_oldStyle)) ] ), ' of ', length( GS ), ' genes found in the Network' ))

Trial.path <- induced_subgraph( net_oldStyle, GS[ GS%in% names(V(net_oldStyle)) ] )

#Trial.path <- E(net_oldStyle)[ V(net_oldStyle)[ names(V(net_oldStyle)) %in% GS] ]
Trial.path <- subgraph.edges( net_oldStyle, E(net_oldStyle)[ .inc(V(net_oldStyle)[GeneSetB])] )

#Remove all CHEBI: nonsense
Trial.path <- subgraph.edges( Trial.path, E(Trial.path)[ grepl( 'CHEBI', edge_attr(Trial.path)$UniqCol)  == F ] )


##_#
foo <- subgraph.edges( net_oldStyle, E(net_oldStyle)[ grepl( 'CHEBI', edge_attr(net_oldStyle)$UniqCol)  == F ] )

#############################################################################################################################
######Cluster the Total Graph
#devtools::install_github("igraph/rigraph")

#Potential Weights For Clustering....
Total$EdgeRep <- 0
Total$Edge <- paste0( Total$from, ':',  Total$to )
key <- table( Total$Edge )
Total$EdgeRep <- key[ Total$Edge ]

Total_Occurence <- rep( 0, length(names(key)) )
names( Total_Occurence ) <- names(key)

SumTally <- function( name,df ){
  Value <- sum( df[ df$Edge == name, ]$Occurance )
  return(c(name,Value))
}

SumTally <- function( name,df ){
  Value <- sum( df[ df$Edge == name, ]$Occurance )
  return(c(name,Value))
}

temp_Total <- Total[ ,c('Edge','Occurance')]
temp_Total$Occurance <- as.numeric(temp_Total$Occurance)

cl <- makeCluster( detectCores()-2 )
registerDoParallel(cl)

L <- length(names(Total_Occurence))
mark<-Sys.time()
#TOs <- foreach( i=names(Total_Occurence)[1:L], .combine=rbind, .export = ls(globalenv()) ) %dopar% SumTally( i,temp_Total )
TOs <- foreach( i=names(Total_Occurence)[1:L], .combine=rbind ) %dopar% SumTally( i,temp_Total )
Sys.time()-mark
# Time difference of 2.260386 days

stopCluster(cl)
#TOs <- read.csv('../ReferenceData/Occurance.csv', row.names=1)
sink_TO <-TOs

#write to file
#write.csv( temp_Total, file='../ReferenceData/Occurance.csv')
#write.csv( TOs, file='../ReferenceData/SumTalliedOccur.csv')
#TOs <- dplyr::distinct(TOs)
row.names(TOs) <- TOs[,1]
TOs <- as.data.frame(TOs)
TOs$V2 <- as.numeric(as.character(TOs$V2))
head(TOs[ Total$Edge, ])

Total$SumOccurence <- TOs[ Total$Edge, ]$V2

Genes <- c( as.character(Total$from), as.character(Total$to) )
Genes <- Genes[ !duplicated(Genes) ]
Genes <- as.data.frame( Genes )

Graph <- list()
for(type in levels(Total$interaction) ){
    eval( parse( text=paste0('Graph$`', type, '` <- graph_from_data_frame(d=Total[ Total$interaction == \'', type, '\', ], vertices=Genes, directed=T) ') ) )
}
#Edge Sum is 1903389
save(Total, file = "~/igraph_Network_Expansion/ReferenceData/Final_Total.RData")

#load("../ReferenceData/Final_Total.tsv")
#Another way to make the Network object But has multiple edges per-Vertex set
net_oldStyle <- graph_from_data_frame(d=Total, vertices=Genes, directed=T) 
Data <- list( DF=Total, NET=net_oldStyle, Gra=Graph )

```

```{r PullCoExp}

#############################################################################################################################
#Add coexpression to the

library("biomaRt")
#dim(Total[ grepl('CHEBI', as.character(Total$UniqCol)) == F, ])
Total <- Total[ grepl('CHEBI', as.character(Total$UniqCol)) == F, ]

mart = useDataset( "hsapiens_gene_ensembl", useEnsembl(biomart="ensembl", version=84) )

Genes <- c( as.character(Total$from), as.character(Total$to) )
Genes <- Genes[ !duplicated(Genes) ]

TRANS <- getBM(attributes=c('hgnc_symbol', 'ensembl_gene_id'), 
      filters = 'hgnc_symbol', 
      values = Genes, 
      mart = mart)

library(dplyr)
Pull_CoExp <- function( Data, SynID, trans, DF){
  #'@Data Character vector of the Brain Region eg. 'DLPFC'
  #'@SynID SynID of the coexpression matrix eg. 'syn21630206'
  #'@trans gene name to ensg dataframe for DF eg. TRANS
  #'@DF The network dataframe eg Total
  
  #Data <- 'DLPFC'
  #SynID <- 'syn21630206'
  #trans <- TRANS
  #DF <- Total[ , colnames(Total) != 'DLPFC_CE' ]
  
  #DLPFC
  CoExp <- read.table(syn_temp$get(SynID)$path, header=T,sep='\t')
  
  #Filter trans for relevant ENSGs:
  trans <- trans[ trans$ensembl_gene_id %in% colnames(CoExp), ]
  
  #Filter CoExp for relevant Genes:
  CoExp_filt <- CoExp[ row.names(CoExp)[ row.names(CoExp) %in% trans$ensembl_gene_id],
                       colnames(CoExp)[ colnames(CoExp) %in% trans$ensembl_gene_id]
                       ]
  #Make the Statistic Column:
  eval(parse(text= paste0( 'DF$', Data, '_CE <- NA' ) ))
  
  #Split DF into unanotatable and need to be annotate:
  Total_UA <- DF[ (as.character(DF$from) %in% trans$hgnc_symbol) == F | (as.character(DF$to) %in% trans$hgnc_symbol) == F, ]
  Total_A <- DF[ (as.character(DF$from) %in% trans$hgnc_symbol) == T & (as.character(DF$to) %in% trans$hgnc_symbol) == T, ]
  
  # Melt the CoExpressed Filtered
  row.names(trans) <- trans$ensembl_gene_id
  row.names(CoExp_filt) <- trans[ row.names(CoExp_filt), ]$hgnc_symbol
  colnames(CoExp_filt) <- trans[ colnames(CoExp_filt), ]$hgnc_symbol
  
  m<-as.matrix(CoExp_filt)
  foo<-data.frame(col = rep(colnames(m), each = nrow(m)), 
             row = rep(rownames(m), ncol(m)), 
             value = as.vector(m))
  
  #Filter the coexpression for the pairs that matter ie not zero
  foo %>% mutate_if(is.factor, as.character) -> foo2
  
  Final_Annote <- foo[ foo$value > 0,]
  Final_Annote %>% mutate_if(is.factor, as.character) -> Final_Annote
  
  temp<-Final_Annote
  temp <- temp[ temp$col != temp$row, ]
  temp <- temp[,c(2,1,3)]
  colnames(temp) <- c( 'col', 'row', 'value')
  
  Exp <- merge( Final_Annote, temp )
  row.names(Exp) <- paste0( Exp$row, ':', Exp$col )
  
  #Split the Total_A (Annotate into zero and non-zero)
  Total_Z <- Total_A[ (Total_A$Edge %in% row.names(Exp)) == F, ]
  eval(parse(text=paste0( 'Total_Z$', Data,'_CE <- 0' ) ))
  
  Total_A <- Total_A[ Total_A$Edge %in% row.names(Exp), ]
  eval(parse(text=paste0( 'Total_A$', Data, '_CE <- Exp[ Total_A$Edge, ]$value') ))
  
  #Merge the datasets
  Total_CE <- dplyr::bind_rows( Total_A,Total_Z )
  DF <- dplyr::bind_rows( Total_CE, Total_UA )
  return(DF)
}

CoExpression <- c( 'syn21630206', 'syn21592277', 
                   'syn21655283', 'syn21965435', 
                   'syn21785669', 'syn21911919', 
                   'syn21898644' )
names( CoExpression ) <- c( 'DLPFC', 'CBE', 
                            'FP', 'IFG', 
                            'PHG', 'STG', 
                            'TCX' )

# Run accross all regions:
for( i in 1:length(CoExpression) ) {
  Total <- Pull_CoExp( names( CoExpression )[i], CoExpression[i], TRANS, Total )
}

sink_Total <- Total

Total$Avg_Cortex_CE <- apply( Total[,c('DLPFC_CE', 'FP_CE', 'IFG_CE', 'PHG_CE', 'STG_CE', 'TCX_CE')], 1, mean )

Total$Avg_All_CE <- apply( Total[,c('DLPFC_CE', 'CBE_CE', 'FP_CE', 'IFG_CE', 'PHG_CE', 'STG_CE', 'TCX_CE')], 1, mean )

sink_Total <- Total
save( Total, file='Total_Net_wCoexpression.RData' )
#save.image('2020_08_13_CoexpAdd.RData')

```

```{r pushToSynapse }

parentId <- 'syn21534582'
activityName = 'Igraph_Full_Networks';
activityDescription = 'Full Protein-Protein Nets';
CODE <- syn_temp$store(synapseclient$Folder(name = "igraphData", parentId = parentId))

thisFileName <- '01_Assemble_Network_Dataframe.Rmd'

# Github link
#TBD_ # thisRepo <- githubr::getRepo(repository = "jgockley62/igraph_Network_Expansion", ref="branch", refName='master')
#TBD_ # thisFile <- githubr::getPermlink(repository = thisRepo, repositoryPath=paste0('code/',thisFileName))

#Set Used SynIDs For Provenance
Syns_Used <- SIFs


# Set annotations
all.annotations = list(
  dataType = 'Network',
  summaryLevel = 'gene',
  assay	 = 'RNAseq',
  tissueTypeAbrv	= c('IFG', 'STG', 'FP', 'PHG', 'TCX', 'DLFPC'), 
  study = c( 'MSBB', 'ROSMAP', 'Mayo' ), 
  organism = 'HomoSapiens',
  consortium	= 'TreatAD',
  genomeAssemblyID = 'GRCh38'
)

# Store  executed = thisFile,
ENRICH_OBJ <-  syn_temp$store( synapseclient$File( path='~/igraph_Network_Expansion/ReferenceData/Final_Total.RData', name = 'Final Total Interaction dataset', parentId=CODE$properties$id ), used = Syns_Used, activityName = activityName, activityDescription = activityDescription)

ENRICH_OBJ <-  syn_temp$store( synapseclient$File( path='~/igraph_Network_Expansion/Total_Net_wCoexpression.RData', name = 'Final Total Interaction dataset With Coexpression', parentId=CODE$properties$id ), used = Syns_Used, activityName = activityName, activityDescription = activityDescription)

```
